feat: Update codeforces_wrapper.py to parse limit value as float instead of int

+++ codeforces_wrapper.py
@@ -21,7 +21,7 @@ def parse_problem(problem_link):
 def split_limit(soup):
     l = soup.split()
     return {
-        "value": int(l[0]),
+        "value": float(l[0]),
         "unit": l[1]
     }
